Activate SelfResolvables last
This change in execution order ensures that constraints, that are created outside of the tree and that need to be resolved by a parent node, are properly activated once its associated view has beed added to the tree. This approach is useful when it's necessary to keep a reference constraint.